+++ cardano-node/src/Cardano/Tracing/Peer.hs
@@ -40,9 +40,9 @@ import qualified Ouroboros.Network.Block as Net
 import qualified Ouroboros.Network.BlockFetch.ClientRegistry as Net
 import           Ouroboros.Network.BlockFetch.ClientState (PeerFetchInFlight (..),
                    PeerFetchStatus (..), readFetchClientState)
+import           Ouroboros.Network.NodeToNode (RemoteAddress)
 
 import           Cardano.Node.Queries
-import Ouroboros.Network.NodeToNode (RemoteAddress)
 
 {- HLINT ignore "Use =<<" -}
 {- HLINT ignore "Use <=<" -}